fix: vfat.io long-poll reload on Android.

+++ apps/mobile/src/core/bridges/useBackgroundBridge.ts
@@ -56,8 +56,12 @@ export function useBackgroundBridges() {
 }
 
 type WebView = import('react-native-webview').WebView;
-type OnLoadStart = import('react-native-webview').WebViewProps['onLoadStart'] &
-  Function;
+type OnLoadStart = (
+  event: Parameters<
+    import('react-native-webview').WebViewProps['onLoadStart'] & Function
+  >[0],
+  treatAsReload?: boolean,
+) => void;
 type OnMessage = import('react-native-webview').WebViewProps['onMessage'] &
   Function;
 
@@ -153,7 +157,7 @@ export function useSetupWebview({
 
   // would be called every time the url changes
   const onLoadStart: OnLoadStart = useCallback(
-    async ({ nativeEvent }) => {
+    async ({ nativeEvent }, treatAsReload = false) => {
       if (
         nativeEvent.url !== urlRef.current &&
         nativeEvent.loading &&
@@ -169,26 +173,28 @@ export function useSetupWebview({
 
       // icon.current = null;
 
-      // Reset the previous bridges
-      backgroundBridgeRefs.current.length &&
-        backgroundBridgeRefs.current.forEach(bridge => {
-          bridge.onDisconnect();
-          sessionService.deleteSession(bridge);
-        });
+      if (treatAsReload) {
+        // Reset the previous bridges
+        backgroundBridgeRefs.current.length &&
+          backgroundBridgeRefs.current.forEach(bridge => {
+            bridge.onDisconnect();
+            sessionService.deleteSession(bridge);
+          });
 
-      // // Cancel loading the page if we detect its a phishing page
-      // const { hostname } = new URL(nativeEvent.url);
-      // if (!isAllowedUrl(hostname)) {
-      //   handleNotAllowedUrl(url);
-      //   return false;
-      // }
+        // // Cancel loading the page if we detect its a phishing page
+        // const { hostname } = new URL(nativeEvent.url);
+        // if (!isAllowedUrl(hostname)) {
+        //   handleNotAllowedUrl(url);
+        //   return false;
+        // }
 
-      const dappOrigin = nativeEvent.url;
-      backgroundBridgeRefs.current = [];
-      const formattedDappOrigin = BUILTIN_SPECIAL_URLS.includes(dappOrigin)
-        ? dappOrigin
-        : urlUtils.canoicalizeDappUrl(dappOrigin).httpOrigin;
-      initializeBackgroundBridge(formattedDappOrigin, true);
+        const dappOrigin = nativeEvent.url;
+        backgroundBridgeRefs.current = [];
+        const formattedDappOrigin = BUILTIN_SPECIAL_URLS.includes(dappOrigin)
+          ? dappOrigin
+          : urlUtils.canoicalizeDappUrl(dappOrigin).httpOrigin;
+        initializeBackgroundBridge(formattedDappOrigin, true);
+      }
     },
     [backgroundBridgeRefs, changeUrl, initializeBackgroundBridge, urlRef],
   );
